performance_locust/requirements.txt
Faker
locust
python-dotenv